# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "API REST para gerenciamento de hotéis, quartos, hóspedes, reservas, pagamentos, cupons e taxas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/hotels/{hotelId}/payments": {
            "get": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Lista os pagamentos de um hotel",
                "description": "Retorna todos os pagamentos vinculados às reservas de um hotel.",
                "operationId": "72dad48cece5255c86343690e5b66927",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "description": "ID do hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pagamentos retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Payment"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para acessar os pagamentos deste hotel"
                    },
                    "404": {
                        "description": "Hotel não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pagamentos"
                ],
                "summary": "Cria um pagamento",
                "description": "Registra um pagamento para uma reserva vinculada ao hotel informado.",
                "operationId": "c0c45ecc0a29b511fd2fd5eaaa09ac95",
                "parameters": [
                    {
                        "name": "hotelId",
                        "in": "path",
                        "description": "ID do hotel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reserve_id",
                                    "payment_method_id",
                                    "value"
                                ],
                                "properties": {
                                    "reserve_id": {
                                        "description": "ID da reserva que receberá o pagamento",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_method_id": {
                                        "description": "ID do método de pagamento",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "value": {
                                        "description": "Valor do pagamento",
                                        "type": "number",
                                        "format": "float",
                                        "example": 250,
                                        "minimum": 0
                                    },
                                    "status": {
                                        "description": "Status inicial do pagamento",
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "external_reference": {
                                        "description": "Referência externa da transação",
                                        "type": "string",
                                        "example": "TXN-123456",
                                        "nullable": true
                                    },
                                    "metadata": {
                                        "description": "Metadados adicionais do pagamento",
                                        "type": "object",
                                        "example": {
                                            "gateway": "stripe",
                                            "transaction_id": "pi_123456"
                                        },
                                        "nullable": true
                                    },
                                    "paid_at": {
                                        "description": "Data e hora de confirmação do pagamento",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-10-06T22:30:00Z",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "reserve_id": 1,
                                "payment_method_id": 1,
                                "value": 250,
                                "status": "pending",
                                "external_reference": "TXN-123456",
                                "metadata": {
                                    "gateway": "stripe",
                                    "transaction_id": "pi_123456"
                                },
                                "paid_at": "2026-10-06T22:30:00Z"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pagamento criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Payment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para criar pagamentos neste hotel"
                    },
                    "404": {
                        "description": "Hotel não encontrado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Payment": {
                "title": "Pagamento",
                "required": [
                    "id",
                    "reserve_id",
                    "payment_method_id",
                    "value",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reserve_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "payment_method_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 250
                    },
                    "status": {
                        "type": "string",
                        "example": "paid",
                        "enum": [
                            "pending",
                            "paid",
                            "refunded",
                            "failed"
                        ]
                    },
                    "external_reference": {
                        "type": "string",
                        "example": "TXN-123456",
                        "nullable": true
                    },
                    "metadata": {
                        "type": "object",
                        "example": {
                            "gateway": "stripe",
                            "transaction_id": "pi_123456"
                        },
                        "nullable": true
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-06T22:30:00.000000Z",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Informe o token no formato: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}